backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
prisma==0.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
sqlalchemy==2.0.25
asyncpg==0.29.0
pytest==7.4.4
httpx==0.26.0